bpo-40268: Move struct _gc_runtime_state to pycore_gc.h (GH-19515)
