+++ Modules/_tracemalloc.c
@@ -1,5 +1,6 @@
 #include "Python.h"
-#include "pycore_pymem.h"
+#include "pycore_gc.h"            // PyGC_Head
+#include "pycore_pymem.h"         // _Py_tracemalloc_config
 #include "pycore_traceback.h"
 #include "hashtable.h"
 #include "frameobject.h"
